Update to latest bukkit release (MC 1.6.4)
This also removes all references to CraftBukkit or MCServer, for full
Bukkit compliance. Hooray!

+++ src/main/java/com/elmakers/mine/bukkit/plugins/magic/spells/GillsSpell.java
@@ -107,7 +107,7 @@ public class GillsSpell extends Spell
 					gills.heal();
 					if (isUnderwater())
 					{
-						int health = player.getHealth();
+						double health = player.getHealth();
 						if (health < 20) 
 						{
 							health = health + healAmount;
